unify router link directive mocks
