+++ src/test/javascript/spec/component/admin/upcoming-exams-and-exercises.component.spec.ts
@@ -6,7 +6,7 @@ import { MockExerciseService } from '../../helpers/mocks/service/mock-exercise.s
 import { ExamManagementService } from 'app/exam/manage/exam-management.service';
 import { MockExamManagementService } from '../../helpers/mocks/service/mock-exam-management.service';
 import { TranslatePipeMock } from '../../helpers/mocks/service/mock-translate.service';
-import { MockRouterLinkDirective } from '../shared/navbar.component.spec';
+import { MockRouterLinkDirective } from '../../helpers/mocks/directive/mock-router-link.directive';
 
 describe('UpcomingExamsAndExercisesComponent', () => {
     let component: UpcomingExamsAndExercisesComponent;
