+++ src/test/javascript/spec/component/overview/course-card.component.spec.ts
@@ -6,7 +6,7 @@ import { ArtemisTestModule } from '../../test.module';
 import { CourseCardComponent } from 'app/overview/course-card.component';
 import { Course } from 'app/entities/course.model';
 import { ArtemisTranslatePipe } from 'app/shared/pipes/artemis-translate.pipe';
-import { MockRouterLinkDirective } from '../lecture-unit/lecture-unit-management.component.spec';
+import { MockRouterLinkDirective } from '../../helpers/mocks/directive/mock-router-link.directive';
 import { SecuredImageComponent } from 'app/shared/image/secured-image.component';
 import { Exercise } from 'app/entities/exercise.model';
 import { MockComponent, MockPipe, MockProvider, MockDirective } from 'ng-mocks';
