+++ src/test/javascript/spec/component/shared/navbar.component.spec.ts
@@ -1,5 +1,4 @@
 import { HttpResponse } from '@angular/common/http';
-import { Directive, Input } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { Router, UrlSerializer } from '@angular/router';
 import { NgbCollapse, NgbDropdown } from '@ng-bootstrap/ng-bootstrap';
@@ -30,21 +29,13 @@ import { MockRouter } from '../../helpers/mocks/mock-router';
 import { MockSyncStorage } from '../../helpers/mocks/service/mock-sync-storage.service';
 import { ArtemisTestModule } from '../../test.module';
 import { OrganizationManagementService } from 'app/admin/organization-management/organization-management.service';
+import {
+    MockRouterLinkActiveOptionsDirective,
+    MockRouterLinkDirective,
+} from '../../helpers/mocks/directive/mock-router-link.directive';
 
 chai.use(sinonChai);
 const expect = chai.expect;
-
-// tslint:disable-next-line:directive-selector
-@Directive({ selector: '[routerLink]' })
-export class MockRouterLinkDirective {
-    @Input('routerLink') data: any;
-}
-
-// tslint:disable-next-line:directive-selector
-@Directive({ selector: '[routerLinkActiveOptions]' })
-export class MockRouterLinkActiveOptionsDirective {
-    @Input('routerLinkActiveOptions') data: any;
-}
 
 class MockBreadcrumb {
     label: string;
@@ -89,8 +80,8 @@ describe('NavbarComponent', () => {
                 MockDirective(NgbDropdown),
                 MockDirective(ActiveMenuDirective),
                 MockDirective(TranslateDirective),
-                MockDirective(MockRouterLinkDirective),
-                MockDirective(MockRouterLinkActiveOptionsDirective),
+                MockRouterLinkDirective,
+                MockRouterLinkActiveOptionsDirective,
                 MockPipe(ArtemisTranslatePipe),
                 MockPipe(FindLanguageFromKeyPipe),
                 MockComponent(NotificationSidebarComponent),
